+++ libminifi/include/io/FileStream.h
@@ -65,7 +65,7 @@ class FileStream : public io::BaseStream {
    */
   void seek(uint64_t offset) override;
 
-  const uint64_t getSize() const override {
+  const size_t getSize() const override {
     return length_;
   }
