+++ libminifi/include/io/NonConvertingStream.h
@@ -171,7 +171,7 @@ class NonConvertingStream : public BaseStream  {
    **/
   int read(uint64_t &value, bool is_little_endian = false) override;
 
-  const uint64_t getSize() const override {
+  const size_t getSize() const override {
       if (composable_stream_ == this) {
         return buffer.size();
       } else {
